+++ packages/web/src/app/shared/viewport-size/viewport-size.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { LayoutModule } from '@angular/cdk/layout';
+import { IfViewportSizeDirective } from './if-viewport-size.directive';
+
+@NgModule({
+  declarations: [IfViewportSizeDirective],
+  imports: [CommonModule, LayoutModule],
+  exports: [IfViewportSizeDirective],
+})
+export class ViewportSizeModule {}